[DDW-622] Move currency related code into a dedicated store
