+++ source/renderer/app/actions/currency-actions.js
@@ -1,0 +1,9 @@
+// @flow
+import Action from './lib/Action';
+
+// ======= CURRENCY ACTIONS =======
+
+export default class CurrencyActions {
+  setCurrencySelected: Action<{ currencyCode: string }> = new Action();
+  toggleCurrencyIsActive: Action<any> = new Action();
+}
